added custom user services for Staff and Manager
