+++ docs/index.html
@@ -78,7 +78,7 @@
                         <div class="col-md-12">
                             <div class="callout callout-info">
                                 <h4>SchemaSpy Analysis of space_renting_db.public</h4>
-                                <p>Generated on Sat Jun 28 14:17 UTC 2025</p>
+                                <p>Generated on Sat Jun 28 14:37 UTC 2025</p>
                             </div>
                         </div>
                     </div>
