+++ docs/info-html.txt
@@ -1,4 +1,4 @@
-date=2025-06-28 14:17:11+0000
+date=2025-06-28 14:37:54+0000
 os=Linux 6.8.0-59-generic
 schemaspy-version=7.0.2
 schemaspy-revision=v7.0.2
