--- src/main/java/com/oussama/space_renting/security/CustomManagerDetailsService.java
@@ -1,37 +1,0 @@
-package com.oussama.space_renting.security;
-
-import com.oussama.space_renting.model.User.User;
-import com.oussama.space_renting.repository.UserRepository;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service("managerDetailsService")
-public class CustomManagerDetailsService implements UserDetailsService {
-    private final UserRepository userRepository;
-
-    public CustomManagerDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    /*
-     * We have to override this function from the UserDetailsService interface
-     * It's used to get user information based on something
-     * In our case email
-     */
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        User user = userRepository.findByEmail( email)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-
-        return new org.springframework.security.core.userdetails.User(
-                user.getEmail(),
-                user.getPassword(),
-                List.of(new SimpleGrantedAuthority("USER"))
-        );
-    }
-}
