+++ src/main/java/com/oussama/space_renting/security/custom_user_details/CustomStaffDetailsService.java
@@ -1,0 +1,33 @@
+package com.oussama.space_renting.security.custom_user_details;
+
+import com.oussama.space_renting.model.Staff.Staff;
+import com.oussama.space_renting.repository.StaffRepository;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service("staffDetailsService")
+public class CustomStaffDetailsService implements UserDetailsService {
+
+    private final StaffRepository staffRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        Staff staff = staffRepository.findByEmail( email)
+                .orElseThrow(() -> new UsernameNotFoundException("Staff not found"));
+
+        return new org.springframework.security.core.userdetails.User(
+                staff.getEmail(),
+                staff.getPassword(),
+                List.of(new SimpleGrantedAuthority("STAFF"))
+        );
+    }
+
+    public CustomStaffDetailsService(StaffRepository staffRepository) {
+        this.staffRepository = staffRepository;
+    }
+}
